:recycle: Move thread table from CPU Module to global kernel resource cache
