+++ Kernel/Src/CPU/SConscript
@@ -29,6 +29,7 @@ sources = [
     build_env.File("Threading/Scheduler.cpp"),
     build_env.File("Threading/Semaphore.cpp"),
     build_env.File("Threading/Spinlock.cpp"),
+    build_env.File("Threading/Thread.cpp"),
     build_env.File("Time/DeltaQueue.cpp"),
     build_env.File("Time/PIT.cpp"),
     build_env.File("Time/Timer.cpp"),
